test(session): move resumed-prefix system message test to part 04
Relocate the test covering resumed transcript prefixes that lack a leading
system message into session_tests_part_04_tests.rs so the session test files
stay within their intended size split. No test logic changed.

Auto-committed-on: macbook
